#350] Code duplication in FEEL libs: Refactored Standard libs to extend BaseStandardFEELLib

+++ dmn-runtime/src/main/java/com/gs/dmn/feel/lib/BaseFEELLib.java
@@ -611,6 +611,6 @@ public abstract class BaseFEELLib<NUMBER, DATE, TIME, DATE_TIME, DURATION> imple
         }
     }
 
-    protected abstract NUMBER valueOf(int number);
+    protected abstract NUMBER valueOf(long number);
     protected abstract int intValue(NUMBER number);
 }
